reorganized component structure
